ci: drop competitive bench regression job, split bench files
The bench/ suite compares TMP against OpenRTB and Cap'n Proto — it exists
to substantiate claims like "37% smaller than OpenRTB", not to guard
internal perf. Running it on every PR via benchstat burned ~20min of CI
measuring third-party libraries we don't own.

Removed the benchmark job entirely. Split bench_test.go into
tmp_bench_test.go (TMP-only marshal/unmarshal) and competitive_bench_test.go
(OpenRTB + the TMP/OpenRTB side-by-side FullExchange pair). Renamed
capnp_bench_test.go to competitive_capnp_test.go for consistency. The
test job still runs `go test ./...` in bench/, so compilation and the
payload-size tests stay covered.
