+++ .github/workflows/ci.yml
@@ -90,148 +90,3 @@ jobs:
           version: v2.11
           working-directory: reference/identity-agent
           args: ./...
-
-  benchmark:
-    name: Benchmark
-    runs-on: ubuntu-latest
-    timeout-minutes: 20
-    if: github.event_name == 'pull_request'
-    permissions:
-      contents: read
-      pull-requests: write
-    steps:
-      - uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
-
-      - uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
-        with:
-          ref: main
-          path: _base
-          fetch-depth: 1
-
-      - uses: actions/setup-go@40f1582b2485089dde7abd97c1529aa768e1baff # v5
-        with:
-          go-version: "1.26"
-
-      - name: Install benchstat
-        run: go install golang.org/x/perf/cmd/benchstat@v0.0.0-20260312031701-16a31bc5fbd0
-
-      - name: Benchmark PR
-        working-directory: bench
-        run: go test -bench=. -benchmem -count=5 ./... | tee /tmp/bench-pr.txt
-
-      - name: Benchmark main
-        working-directory: _base/bench
-        run: go test -bench=. -benchmem -count=5 ./... | tee /tmp/bench-main.txt
-
-      - name: Compare benchmarks
-        run: |
-          benchstat /tmp/bench-main.txt /tmp/bench-pr.txt | tee /tmp/benchstat.txt
-
-      - name: Check for regressions
-        id: check_regressions
-        run: |
-          # Parse benchstat output for statistically significant regressions > 5%.
-          # benchstat lines look like:  BenchmarkName  1.234µ ± ∞ ¹   1.345µ ± ∞ ¹  +9.00% (p=0.008 n=5)
-          # We match lines with a positive % change and p < 0.05.
-          regressions=""
-          while IFS= read -r line; do
-            # Extract percentage and p-value from lines showing a change
-            if echo "$line" | grep -qE '\+[0-9]+\.[0-9]+% \(p='; then
-              pct=$(echo "$line" | sed -n 's/.*+\([0-9][0-9]*\.[0-9][0-9]*\)%.*/\1/p')
-              pval=$(echo "$line" | sed -n 's/.*p=\([0-9][0-9]*\.[0-9][0-9]*\).*/\1/p')
-              name=$(echo "$line" | awk '{print $1}')
-              # Validate extracted values are numeric before using in awk
-              if ! [[ "$pct" =~ ^[0-9]+\.[0-9]+$ ]] || ! [[ "$pval" =~ ^[0-9]+\.[0-9]+$ ]]; then
-                continue
-              fi
-              # Check: regression > 5% AND statistically significant (p < 0.05)
-              if awk -v p="$pct" -v pv="$pval" 'BEGIN{exit !(p > 5.0 && pv < 0.05)}'; then
-                regressions="${regressions}${name}: +${pct}% (p=${pval})"$'\n'
-              fi
-            fi
-          done < /tmp/benchstat.txt
-
-          if [ -n "$regressions" ]; then
-            echo "found=true" >> "$GITHUB_OUTPUT"
-            printf '%s' "$regressions" | tee /tmp/regressions.txt
-            echo ""
-            echo "::error::Benchmark regressions detected (>5%, p<0.05):"
-            printf '%s' "$regressions"
-          else
-            echo "found=false" >> "$GITHUB_OUTPUT"
-            echo "No significant regressions found."
-          fi
-
-      - name: Remove stale regression comment
-        if: steps.check_regressions.outputs.found == 'false'
-        uses: actions/github-script@f28e40c7f34bde8b3046d885e986cb6290c5673b # v7
-        with:
-          script: |
-            const { data: comments } = await github.rest.issues.listComments({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              issue_number: context.issue.number,
-              per_page: 100,
-            });
-            const isBotComment = c => c.user.login === 'github-actions[bot]' && c.body.startsWith('## Benchmark Regression');
-            const existing = comments.find(isBotComment);
-            if (existing) {
-              await github.rest.issues.deleteComment({
-                owner: context.repo.owner,
-                repo: context.repo.repo,
-                comment_id: existing.id,
-              });
-            }
-
-      - name: Comment on regression
-        if: steps.check_regressions.outputs.found == 'true'
-        uses: actions/github-script@f28e40c7f34bde8b3046d885e986cb6290c5673b # v7
-        with:
-          script: |
-            const fs = require('fs');
-            const regressions = fs.readFileSync('/tmp/regressions.txt', 'utf8').trim().replace(/`/g, '\\`');
-            const full = fs.readFileSync('/tmp/benchstat.txt', 'utf8').replace(/`/g, '\\`');
-            const body = [
-              '## Benchmark Regression Detected',
-              '',
-              '```',
-              regressions,
-              '```',
-              '',
-              '<details><summary>Full benchstat output</summary>',
-              '',
-              '```',
-              full,
-              '```',
-              '',
-              '</details>',
-            ].join('\n');
-            const { data: comments } = await github.rest.issues.listComments({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              issue_number: context.issue.number,
-              per_page: 100,
-            });
-            const isBotComment = c => c.user.login === 'github-actions[bot]' && c.body.startsWith('## Benchmark Regression');
-            const existing = comments.find(isBotComment);
-            if (existing) {
-              await github.rest.issues.updateComment({
-                owner: context.repo.owner,
-                repo: context.repo.repo,
-                comment_id: existing.id,
-                body,
-              });
-            } else {
-              await github.rest.issues.createComment({
-                owner: context.repo.owner,
-                repo: context.repo.repo,
-                issue_number: context.issue.number,
-                body,
-              });
-            }
-
-      - name: Fail on regression
-        if: steps.check_regressions.outputs.found == 'true'
-        run: |
-          echo "Benchmark regressions exceed 5% threshold. See PR comment for details."
-          exit 1
