+++ bench/competitive_bench_test.go
@@ -6,7 +6,9 @@ import (
 	"testing"
 )
 
-// --- Payload Size Comparison ---
+// Competitive benchmarks comparing TMP against OpenRTB. Run manually (not in CI).
+//   go test -bench=. -benchmem ./bench/...
+// See also competitive_capnp_test.go for TMP vs Cap'n Proto.
 
 func TestPayloadSizes(t *testing.T) {
 	ortbReq, _ := json.Marshal(RealisticOpenRTBRequest())
@@ -35,8 +37,6 @@ func TestPayloadSizes(t *testing.T) {
 	fmt.Printf("  TMP is %.0f%% smaller than OpenRTB\n", (1-float64(tmpTotal)/float64(ortbTotal))*100)
 	fmt.Println()
 }
-
-// --- OpenRTB Serialization Benchmarks ---
 
 func BenchmarkOpenRTB_Request_Marshal(b *testing.B) {
 	req := RealisticOpenRTBRequest()
@@ -71,114 +71,16 @@ func BenchmarkOpenRTB_RoundTrip(b *testing.B) {
 	}
 }
 
-// --- TMP Context Match Benchmarks ---
-
-func BenchmarkTMP_ContextRequest_Marshal(b *testing.B) {
-	req := RealisticTMPContextRequest()
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		data, _ := json.Marshal(req)
-		b.SetBytes(int64(len(data)))
-	}
-}
-
-func BenchmarkTMP_ContextRequest_Unmarshal(b *testing.B) {
-	data, _ := json.Marshal(RealisticTMPContextRequest())
-	b.SetBytes(int64(len(data)))
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		var req TMPContextRequest
-		json.Unmarshal(data, &req)
-	}
-}
-
-func BenchmarkTMP_ContextResponse_Marshal(b *testing.B) {
-	resp := RealisticTMPContextResponse()
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		data, _ := json.Marshal(resp)
-		b.SetBytes(int64(len(data)))
-	}
-}
-
-// --- TMP Identity Match Benchmarks ---
-
-func BenchmarkTMP_IdentityRequest_Marshal(b *testing.B) {
-	req := RealisticTMPIdentityRequest()
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		data, _ := json.Marshal(req)
-		b.SetBytes(int64(len(data)))
-	}
-}
-
-func BenchmarkTMP_IdentityRequest_Unmarshal(b *testing.B) {
-	data, _ := json.Marshal(RealisticTMPIdentityRequest())
-	b.SetBytes(int64(len(data)))
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		var req TMPIdentityRequest
-		json.Unmarshal(data, &req)
-	}
-}
-
-func BenchmarkTMP_IdentityResponse_Marshal(b *testing.B) {
-	resp := RealisticTMPIdentityResponse()
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		data, _ := json.Marshal(resp)
-		b.SetBytes(int64(len(data)))
-	}
-}
-
-// --- Full Round-Trip Comparisons ---
-
-func BenchmarkTMP_Context_RoundTrip(b *testing.B) {
-	req := RealisticTMPContextRequest()
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		data, _ := json.Marshal(req)
-		var got TMPContextRequest
-		json.Unmarshal(data, &got)
-		b.SetBytes(int64(len(data)))
-	}
-}
-
-func BenchmarkTMP_Identity_RoundTrip(b *testing.B) {
-	req := RealisticTMPIdentityRequest()
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		data, _ := json.Marshal(req)
-		var got TMPIdentityRequest
-		json.Unmarshal(data, &got)
-		b.SetBytes(int64(len(data)))
-	}
-}
-
-// --- Combined: Full TMP vs Full OpenRTB ---
-
 func BenchmarkOpenRTB_FullExchange(b *testing.B) {
 	req := RealisticOpenRTBRequest()
 	respFixture := RealisticOpenRTBResponse()
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		// SSP marshals request
 		reqData, _ := json.Marshal(req)
-		// DSP unmarshals request
 		var gotReq BidRequest
 		json.Unmarshal(reqData, &gotReq)
-		// DSP marshals response
 		respData, _ := json.Marshal(respFixture)
-		// SSP unmarshals response
 		var gotResp BidResponse
 		json.Unmarshal(respData, &gotResp)
 		b.SetBytes(int64(len(reqData) + len(respData)))
@@ -193,7 +95,6 @@ func BenchmarkTMP_FullExchange(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		// Context match: marshal request, unmarshal, marshal response, unmarshal
 		ctxReqData, _ := json.Marshal(ctxReq)
 		var gotCtxReq TMPContextRequest
 		json.Unmarshal(ctxReqData, &gotCtxReq)
@@ -201,7 +102,6 @@ func BenchmarkTMP_FullExchange(b *testing.B) {
 		var gotCtxResp TMPContextResponse
 		json.Unmarshal(ctxRespData, &gotCtxResp)
 
-		// Identity match: same
 		idReqData, _ := json.Marshal(idReq)
 		var gotIdReq TMPIdentityRequest
 		json.Unmarshal(idReqData, &gotIdReq)
